Refactor wagtail admin
