+++ example/tests/admin/test_admin_edit_handlers.py
@@ -5,15 +5,15 @@ import pytest
 from django.utils.formats import date_format
 
 from salesman.admin import utils
-from salesman.admin.edit_handlers import (
+from salesman.admin.mixins import OrderAdminMixin
+from salesman.admin.wagtail.edit_handlers import (
     OrderAdminPanel,
     OrderCheckboxPanel,
     OrderDatePanel,
     OrderItemsPanel,
     ReadOnlyPanel,
 )
-from salesman.admin.forms import WagtailOrderModelForm
-from salesman.admin.mixins import OrderAdminMixin
+from salesman.admin.wagtail.forms import WagtailOrderModelForm
 from salesman.admin.wagtail_hooks import OrderAdmin
 from salesman.core.utils import get_salesman_model
