+++ salesman/admin/wagtail/forms.py
@@ -1,0 +1,7 @@
+from wagtail.admin.forms import WagtailAdminModelForm
+
+from ..forms import OrderModelForm
+
+
+class WagtailOrderModelForm(OrderModelForm, WagtailAdminModelForm):
+    pass
