+++ salesman/admin/wagtail/helpers.py
